+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/abc/ScriptPack.java
@@ -247,9 +247,9 @@ public class ScriptPack extends AS3ClassTreeItem {
 
             Trait trait = traits.get(t);
 
-            if (!(trait instanceof TraitClass)) {
+            //if (!(trait instanceof TraitClass)) {
                 writer.startTrait(t);
-            }
+            //}
             Multiname name = trait.getName(abc);
             Namespace ns = name.getNamespace(abc.constants);
             if ((ns.kind == Namespace.KIND_PACKAGE) || (ns.kind == Namespace.KIND_PACKAGE_INTERNAL)) {
@@ -257,7 +257,8 @@ public class ScriptPack extends AS3ClassTreeItem {
             } else {
                 trait.toString(abcIndex, null, convertData, "", abc, false, exportMode, scriptIndex, -1, writer, new ArrayList<>(), parallel, false);
             }
-            if (!(trait instanceof TraitClass)) {
+            if (!(trait instanceof TraitClass)) 
+            {
                 writer.endTrait();
             }
             first = false;